pyproject.toml
[project]
name = "auth-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1,<6.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg-binary>=3.3.2",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
